Examples and more Board functionality
Added two files in the examples folder. Game.java is what I'm using to do test runs of games. StupidAI.java is a very simple implementation of Player. It just chooses moves randomly (which of course is not the best by a long shot).

I also added some methods to Board, all of which are implemented, so no additional abstract methods to worry about! They're mostly to make output nicer and more informative while playing the game.